Updated calendar.view
Fixed the getColor function so it now returns the correct color codes (meaning the color scale can be implemented, i have in place a temporary color scale right now).

also i have a semi working virtual backend for calendar events so the app remembers what days have events (and marks them accordingly) but the app does not remember the value of the event (ie the mood) so every time the user leaves the tracking view the mood get resets to null. working on this next.